move get_model_arch to archs.py
